+++ packages/projection-lifecycle/src/index.ts
@@ -1,3 +1,5 @@
 export * from "./lease.js";
 export * from "./materialize.js";
+export * from "./wal.js";
+export * from "./append.js";
 export * from "./lifecycle.js";